feat(CountdownTimer): Test added to check expire message
